requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

# PostgreSQL Database Layer (Section 3.1.2 B)
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Text2SQL Engine (Section 3.1.3 C)
google-generativeai>=0.3.0
sqlparse>=0.4.4

# API Layer (Bonus Feature)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
